Driving Car
Hi, this is a code in Unity C# for a car driving on a road in loops which waits about 9 seconds until it teleports on the beginning of the road and starts again. This code works fine but I am totally a beginner so I am sure I can make this code way simpler. Do you have any idea?